refactor: rename file from Thumb to Thumbs
